move logic in BaseTextImageEnrichmentModel
